# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorHub - Multivendor Social Media, Sales & Inventory Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SOCIAL ACCOUNTS ──────────────────────────────────────────────────────
  /social-accounts:
    get:
      operationId: listSocialAccounts
      tags: [social-accounts]
      summary: List connected social accounts
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of social accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialAccount"
    post:
      operationId: createSocialAccount
      tags: [social-accounts]
      summary: Connect a social media account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialAccountInput"
      responses:
        "201":
          description: Connected account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/{id}:
    get:
      operationId: getSocialAccount
      tags: [social-accounts]
      summary: Get a social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Social account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
    delete:
      operationId: deleteSocialAccount
      tags: [social-accounts]
      summary: Disconnect a social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected

  # ─── POSTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, industry, status, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
        status:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        clerkUserId:
          type: ["string", "null"]
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name, industry, email]
      properties:
        name:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        address:
          type: string
        logoUrl:
          type: string
        description:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        website:
          type: string
        address:
          type: string
        logoUrl:
          type: string
        description:
          type: string
        status:
          type: string
...
    SocialAccount:
      type: object
      required: [id, vendorId, platform, accountName, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        platform:
          type: string
        accountName:
          type: string
        accountId:
          type: ["string", "null"]
        profileUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
        followersCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    SocialAccountInput:
      type: object
      required: [vendorId, platform, accountName]
      properties:
        vendorId:
          type: integer
        platform:
          type: string
        accountName:
          type: string
        accountId:
          type: string
        profileUrl:
          type: string
        avatarUrl:
          type: string
        followersCount:
          type: integer

    # ─── POST ───────────────────────────────────────────────────────────────
...